Extract parse as util
