+++ ui/packages/components/src/InvokeButton/InvokeModal.tsx
@@ -1,8 +1,10 @@
 import { useState } from 'react';
 import { Alert } from '@inngest/components/Alert';
-import { Button, NewButton } from '@inngest/components/Button';
+import { NewButton } from '@inngest/components/Button';
 import { CodeBlock } from '@inngest/components/CodeBlock';
 import { Modal } from '@inngest/components/Modal';
+
+import { parseCode } from './utils';
 
 const initialCode = JSON.stringify({ data: {} }, null, 2);
 
@@ -90,50 +92,4 @@ export function InvokeModal({ doesFunctionAcceptPayload, isOpen, onCancel, onCon
       </form>
     </Modal>
   );
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return typeof value === 'object' && value !== null && !Array.isArray(value);
-}
-
-function parseCode(code: string): {
-  data: Record<string, unknown>;
-  user: Record<string, unknown> | null;
-} {
-  if (typeof code !== 'string') {
-    throw new Error("The payload form field isn't a string");
-  }
-
-  let payload: Record<string, unknown>;
-  const parsed: unknown = JSON.parse(code);
-  if (!isRecord(parsed)) {
-    throw new Error('Parsed JSON is not an object');
-  }
-
-  payload = parsed;
-
-  let { data } = payload;
-  if (data === null) {
-    data = {};
-  }
-  if (!isRecord(data)) {
-    throw new Error('The "data" field must be an object or null');
-  }
-
-  let user: Record<string, unknown> | null = null;
-  if (payload.user) {
-    if (!isRecord(payload.user)) {
-      throw new Error('The "user" field must be an object or null');
-    }
-    user = payload.user;
-  }
-
-  const supportedKeys = ['data', 'user'];
-  for (const key of Object.keys(payload)) {
-    if (!supportedKeys.includes(key)) {
-      throw new Error(`Property "${key}" is not supported when invoking a function`);
-    }
-  }
-
-  return { data, user };
 }
